+++ src/components/MARCTable/MARCTableRowContainer.js
@@ -27,6 +27,9 @@ export const MARCTableRowContainer = ({
           order={data.order}
           action={data.action}
           subaction={data.field?.subfields?.[0].subaction}
+          field={data.field?.subfields?.[0].field}
+          indicator1={data.field?.subfields?.[0].indicator1}
+          indicator2={data.field?.subfields?.[0].indicator2}
           columnWidths={columnWidths}
           isFirst={i === 0}
           isLast={i === (fields.length - 1)}
@@ -43,6 +46,9 @@ export const MARCTableRowContainer = ({
                 order={data.order}
                 action={data.action}
                 subaction={subfield.subaction}
+                field={subfield.field}
+                indicator1={subfield.indicator1}
+                indicator2={subfield.indicator2}
                 columnWidths={columnWidths}
                 isSubline
                 subfieldIndex={idx}