+++ test/bigtest/interactors/marc-table-interactor.js
@@ -39,6 +39,7 @@ class MARCTableRowInteractor {
   indicator2 = scoped('[data-test-marc-table-indicator2]', TextFielInteractor);
   subfield = scoped('[data-test-marc-table-subfield]', TextFielInteractor);
   subaction = scoped('[data-test-marc-table-subaction]', SelectInteractor);
+  dataTextField = scoped('[data-test-marc-table-data-text]', TextAreaInteractor);
   dataFindField = scoped('[data-test-marc-table-data-find]', TextAreaInteractor);
   dataReplaceField = scoped('[data-test-marc-table-data-replace]', TextAreaInteractor);
   position = scoped('[data-test-marc-table-position]', SelectInteractor);